+++ src/Login/Cookie/CookieLoginMiddleware.php
@@ -131,7 +131,7 @@ final class CookieLoginMiddleware implements MiddlewareInterface
             return;
         }
 
-        if ($expires < time()) {
+        if ($expires !== 0 && $expires < time()) {
             $this->logger->warning('Unable to authenticate user by cookie. Lifetime has expired.');
             return;
         }