+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Chg #71: Add token logging when login was failed (@xepozz)
 - Chg #58: Raise the minimum version of PHP to 8.0 and did refactoring using the features of it (@xepozz, @rustamwin)
 - Chg #58: Raise version of `yiisoft/access` to `^2.0` (@rustamwin)
+- Enh #83: Allow to create a session cookie via `CookieLogin` (@rustamwin)
 
 ## 2.0.0 February 15, 2023